practice : useRef의 사용
- 특정 DOM 선택
- 컴포넌트 외에서 조회 및 수정 가능한 변수 관리
