+++ src/UserList.js
@@ -8,29 +8,11 @@ function User({ user }) {
   );
 }
 
-function UserList() {
-  const users = [
-    {
-      id: 1,
-      username: "velopert",
-      email: "public.velopert@gmail.com",
-    },
-    {
-      id: 2,
-      username: "tester",
-      email: "tester@example.com",
-    },
-    {
-      id: 3,
-      username: "liz",
-      email: "liz@example.com",
-    },
-  ];
-
+function UserList({ users }) {
   return (
     <div>
-      {users.map((user, index) => (
-        <User user={user} key={index} /> // users 매핑하는 과정에서 key 값 필요
+      {users.map((user) => (
+        <User user={user} key={user.id} />
       ))}
     </div>
   );
